requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
filelock==3.13.1
pillow==10.3.0
platformdirs==4.2.0
sqlparse==0.5.0
virtualenv==20.25.0
